+++ packages/patrol_cli/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix "error while disposing" when SIGINT (CTRL+C) is received when waiting for
   app to start on physical iOS device (#1073)
+- Fix `patrol build` crashing (#1074)
 
 ## 1.0.8